Replace PaymentLinkResource.cs to PaymentLinkResponse.cs (#616)
